+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.3.4 - 2019-05-05
+
+### Fixed
+
+- Bug where ls from nested path with no arg was broken
+
+## 0.3.3 - 2019-05-05
+
+### Changed
+
+- Updated UI for history container
+
 ## 0.3.2 - 2019-04-30
 
 ### Fixed